// src/interfaces/mainCampus.d.ts
export interface I_Main_Campus {
    _id: string,
    campusName: string,
    schoolName: string,
    email: string,
    mobileNo: string,
    currency: string,
    currencySymbol: string,
    city: string,
    state: string,
    address: string,
    createdAt: string,
    updatedAt: date | undefined,
    deletedAt: date | undefined
}